backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-dotenv==1.4.2
django-filter==25.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.0
